add generation of wrapper python scripts
The flux/wrapper directory holds a custom wrapper for each .so
library type, and we can use it (with the base Wrapper as the
parent class) to more quickly provide an interface to
existing functions.
