+++ src/bindings/python/_flux/Makefile.am
@@ -41,6 +41,11 @@ fluxpyso_CODEGENINPUT = \
 	_idset.so \
 	_hostlist.so
 
+fluxpyso_WRAPPERS = \
+	$(top_builddir)/src/bindings/python/flux/wrapper/core.py \
+	$(top_builddir)/src/bindings/python/flux/wrapper/idset.py \
+	$(top_builddir)/src/bindings/python/flux/wrapper/hostlist.py
+
 fluxpyso_PYTHON = \
 	__init__.py
 
@@ -152,6 +157,9 @@ fluxpyso_CODEGENINPUT += \
 fluxpyso_CODEGENJSON += \
 	codegen/_security.json
 
+fluxpyso_WRAPPERS += \
+	$(top_builddir)/src/bindings/python/flux/wrapper/security.py
+
 nodist_fluxbindinginclude_HEADERS +=\
 	_security_preproc.h
 
@@ -191,10 +199,13 @@ lib-copy-vpath: ${fluxpyso_PYTHON} ${fluxpyso_LTLIBRARIES}
 		$(LN_S) .libs/$$LIB ./ $(STDERR_DEVNULL) || true; \
 	done
 
-all-local: lib-copy-vpath ${fluxpyso_CODEGENJSON}
+all-local: lib-copy-vpath ${fluxpyso_CODEGENJSON} ${fluxpyso_WRAPPERS}
 
 ${fluxpyso_CODEGENJSON}:
-	$(AM_V_GEN)$(PYTHON) codegen.py generate --out $(srcdir)/codegen ${fluxpyso_CODEGENINPUT}
+	$(AM_V_GEN)$(PYTHON) $(srcdir)/codegen.py dump --out $(srcdir)/codegen ${fluxpyso_CODEGENINPUT}
+
+${fluxpyso_WRAPPERS}: ${fluxpyso_CODEGENJSON}
+	$(AM_V_GEN)$(PYTHON) $(srcdir)/codegen.py generate --out $(top_srcdir)/src/bindings/python/flux/wrapper ${fluxpyso_CODEGENJSON}
 
 # Support VPATH builds
 clean-local-vpath:
@@ -202,7 +213,11 @@ clean-local-vpath:
 
 clean-local: clean-local-vpath
 	-rm -f *.c *.so *.pyc *.pyo *_clean.h *_preproc.h
-	-rm -rf __pycache__
+	-rm -rf __pycache__ ./codegen/*.json
+	-rm -rf $(top_builddir)/src/bindings/python/flux/wrapper/hostlist.py
+	-rm -rf $(top_builddir)/src/bindings/python/flux/wrapper/core.py
+	-rm -rf $(top_builddir)/src/bindings/python/flux/wrapper/security.py
+	-rm -rf $(top_builddir)/src/bindings/python/flux/wrapper/idset.py
 
 install-data-hook:
 	$(AM_V_at)echo Linking python modules in non-standard location... && \
